Prevent self-assignment as a child hotkey (#76)
* Prevent self-assignment as a child hotkey

Add a check to prevent users from setting their own wallet's hotkey as a child. Also, update the `print_all_takes` function to include the SS58 address in the table title for better clarity.

* Enhance table title readability with color formatting

This commit updates the table title in the 'print_all_takes' function to include bright magenta color for the ss58 identifier. This change improves readability and visual distinction of the table header in the CLI output.